requirements.txt
flask
pymongo[srv]
flask_cors
requests==2.24.0
urllib3==1.25.10 
bs4==0.0.1
pandas
python-dotenv
gunicorn==20.0.4
bson